cloud_functions/caption/requirements.txt
openai>=1.12.0
requests>=2.31.0
firebase-admin>=6.4.0
google-cloud-firestore>=2.15.0
google-cloud-storage>=2.15.0
google-cloud-logging>=3.10.0
